Use notes to store p4 info
P4 info is now stored in git notes. This doesn't modify the commit message
when submitting to Perforce which makes it easier possible to share the git
repo.

Note that you need to re-import from Perforce since it's not backwards
compatible. Notes in git are relatively new, so you need a fairly recent
git version (>= 1.7, although 1.6.6 might also work).

Change-Id: Ia058c3b80060059dfeb9cc27d1b17047c4078707