Add ability to dump import to a file
Added --file-dump
This option toggles whether git-p4 will automatically try to import
the Perforce code into a git repository or whether it will dump to a
file for the user to do manually.

git-p4 was continually crashing for me with out of memory errors
on a large repository with a couple years or history. I attempted
to modify it to write the data in pieces into the fast-import process,
but that still caused memory errors. There is probably a way to fix
that problem, but this method allows you to import large repositories
and avoids having to reprocess from Perforce if the fast-import fails.