Pipe fast-import data directly to git.
This prevents out of memory issues from occurring on large repositories
when we try to buffer all of the dat using cStringIO.  It is also a
little faster because we can run fast-import while waiting for the p4
server.

LargeFileWriter now doesn't store all the data.  The debug data can be
written to any file like object, this is used to implement the debug
data dumping.

Exception handling for fast-import has been moved to enclose all of the
fast-import code and be in one place.  We no longer bother to read
stderr from git and output it on our own stderr, instead we just let git
inherit our stderr.  This ensures all the error output will get printed.

Note: most of this change is indentation changes, you can pass
--ignore-space-change or similar to git diff to see the "real" changes.