Add new abort option to submit
When the user doesn't edit the submit template we bring up a question
"Submit anyway". This patch renames the "No" otpion to "Abort" which
doesn't submit the current (and all subsequent) changes, leaving the
repo in the current state.

The "No" option available in older versions seems to be too
dangerous and confusing so it got removed: choosing "No" used to
skip the current patch, but still updated the pointers. This causes
the git and p4 repos to be out of sync, making it very likely that
subsequent commits will fail. If it is necessary to skip a patch
that is in the git repo, this can be achieved by doing "git update-ref
refs/remotes/p4/master <sha1>".

This fixes issue #5