Add new parameter --no-getuserlist to sync/rebase command
When the parameter --no-getuserlist is passed to the git-p4 sync or
rebase command, the cached userlist (~/.gitp4-usercache.txt) will
be used even when an author of a commit can't be find in the cached
list. Normally the userlist is retrieved from Perforce when an
author can't be found. The use of the parameter allows the use of
hand-tailored user lists that contain historical users that are no
longer in the Perforce user database.