Add new option --fuzzy-tags
When the files included by the view spec of a label doesn't match
the files really included in the label, we normally don't create
a tag.

When the option --fuzzy-tags is specified on the command line all
files and revisions included in the label are compared with the
files and revisions in the repository. If all files match we still
create the label. In other words, we create a label if the files
included in the label is a subset of the files in the repository.