Add filters
This change adds three new parameters to the sync (and clone)
command: --tree-filter, --msg-filter, and --content-filter.
All three take the path to a filter script. Note that the path
should be an absolute path for the scripts to work correctly.

These filter scripts allow to apply tree, log message, and/or
content filters during the import of a change.

A tree filter is a script that gets the path of one file (in p4
notation, e.g. //depot/branch/file.txt) on stdin. It can output
a modified path on stdout, or return an empty string to ignore
this file.

The msg filter is a script that gets the commit message on stdin
and returns a modified commit message.

The content filter gets called for text files and is passed the
name of a temporary file on stdin and can modify that file.

These filters are useful when importing an entire p4 repo. They
allow to rearrange the location of some files (or ignore files
that accidentally got submitted in p4). The content filter is
useful to do whitespace cleanup or remove/replace some file
content that should not be in a repo, e.g. passwords.