Don't write "from" lines
This fixes a problem when importing branches: if the new branch in
Perforce is a subset of the parent branch the import didn't work as
expected: the entire parent branch got imported into git. The reason
is the "from" line.

This means that we shouldn't write the "from" line. According to git
documentation these lines can be omitted. In that case it will take
the previous commit of the current branch. Since Perforce only allows
linear history this should be sufficient.

If we start a new branch we list all the files included in that branch,
so we can't add "from" if we want the trees in Perforce and git to match
(otherwise git starts out with the "from" tree instead of an empty tree).